+++ packages/cli/src/commands.ts
@@ -7,9 +7,22 @@ import docgen from './docgen/handler';
 import docs from './docs/handler';
 import { clean, install, pwd, list } from './repo/handler';
 import expandAdaptors from './util/expand-adaptors';
+import printVersions from './util/print-versions';
+
+type CommandList =
+  | 'execute'
+  | 'compile'
+  | 'repo-clean'
+  | 'repo-install'
+  | 'repo-install'
+  | 'repo-pwd'
+  | 'version'
+  | 'docs'
+  | 'docgen'
+  | 'test';
 
 export type Opts = {
-  command?: string;
+  command?: CommandList;
 
   adaptor?: boolean | string;
   adaptors?: string[];
@@ -27,6 +40,7 @@ export type Opts = {
   packages?: string[];
   specifier?: string; // docgen
   repoDir?: string;
+  timeout?: number; // ms
   statePath?: string;
   stateStdin?: string;
 };
@@ -41,10 +55,12 @@ const handlers = {
   ['repo-install']: install,
   ['repo-pwd']: pwd,
   ['repo-list']: list,
+  version: async (_opts: SafeOpts, logger: Logger) => printVersions(logger),
 };
 
-export type SafeOpts = Required<Omit<Opts, 'log'>> & {
+export type SafeOpts = Required<Omit<Opts, 'log' | 'adaptor'>> & {
   log: Record<string, LogLevel>;
+  adaptor: string | boolean;
 };
 
 // Top level command parser
@@ -52,12 +68,18 @@ const parse = async (basePath: string, options: Opts, log?: Logger) => {
   const opts = ensureOpts(basePath, options);
   const logger = log || createLogger(CLI, opts);
 
+  // A bit janky but in execute and test, always print version info FIRST
+  // Should we ALwAYS just do this? It logs to info so you wouldn't usually see it on eg test, docs
+  if (opts.command === 'execute' || opts.command === 'test') {
+    await printVersions(logger);
+  }
+
   if (opts.adaptors && opts.expand) {
     // Note that we can't do this in ensureOpts because we don't have a logger configured yet
     opts.adaptors = expandAdaptors(opts.adaptors, logger);
   }
 
-  if (opts.command! == 'test' && !opts.repoDir) {
+  if (/^(test|version)$/.test(opts.command) && !opts.repoDir) {
     logger.warn(
       'WARNING: no repo module dir found! Using the default (/tmp/repo)'
     );
@@ -75,6 +97,7 @@ const parse = async (basePath: string, options: Opts, log?: Logger) => {
     process.exit(1);
   }
 
+  // @ts-ignore types on SafeOpts are too contradictory for ts, see #115
   return handler(opts, logger);
 };