+++ packages/cli/src/process/spawn.ts
@@ -5,9 +5,10 @@
 import path from 'node:path';
 import * as url from 'url';
 import { fork } from 'node:child_process';
+import process from 'node:process';
 import type { Opts } from '../commands';
 
-type Messages =  { done?: boolean, init?: boolean};
+type Messages = { done?: boolean; init?: boolean; exitCode?: number };
 
 // The default export will create a new child process which calls itself
 export default function (basePath: string, opts: Opts) {
@@ -17,22 +18,23 @@ export default function (basePath: string, opts: Opts) {
 
     // Allows us to load an ESM module from a text string
     '--experimental-vm-modules',
-
-    // Allows us to do import('path/to/language-common') in the linker
-    '--experimental-specifier-resolution=node',
   ];
 
   const dirname = path.dirname(url.fileURLToPath(import.meta.url));
-  
   const child = fork(`${dirname}/process/runner.js`, [], { execArgv });
 
-  child.on('message', ({ done, init }: Messages) => {
+  child.on('message', ({ done, init, exitCode }: Messages) => {
     if (init) {
       child.send({ init: true, basePath, opts });
     }
+
     if (done) {
       child.kill();
-      process.exit(0);
+      process.exit(exitCode);
     }
+  });
+
+  child.on('close', (code: number) => {
+    process.exitCode = code;
   });
 }